% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim.R
\name{buildCorMatrix}
\alias{buildCorMatrix}
\title{Build a correlation matrix of a given dimension and given correlation
that is the same for each pair of variabes}
\usage{
buildCorMatrix(dimOfMat, corrOfVar = 0.5)
}
\arguments{
\item{dimOfMat}{Matrix dimension}

\item{corrOfVar}{Correlation}
}
\value{
Correlation matrix
}
\description{
Build a correlation matrix of a given dimension and given correlation
that is the same for each pair of variabes
}
