% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim.R
\name{buildNormSamp}
\alias{buildNormSamp}
\title{Build a matrix whose columns are samples of standard normally distributed
variables with given correlations}
\usage{
buildNormSamp(corMatrix, numOfSamp = 1000)
}
\arguments{
\item{corMatrix}{Correlation matrix of variables that gives the number of variable as well}

\item{numOfSamp}{Number of samples of each variable. Default is 1000}
}
\value{
Matrix where columns are samples of variables
}
\description{
Build a matrix whose columns are samples of standard normally distributed
variables with given correlations
}
