\name{FSim-package}
\alias{FSim-package}
\alias{FSim}
\docType{package}
\title{
A package to search functionally similar genes
}
\description{
The package can be used to search functionally similar genes by given a
gene, a group of GO terms or a list of keywords.
}
\details{
\tabular{ll}{
Package: \tab FSim\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2012-10-27\cr
License: \tab GPL\cr
}
SearchGene(gid="10", an.go=an.Hs.egGO, n=5)
}
\author{
Qiang Hu
}
\references{}
\keyword{ package }
\seealso{}
\examples{
#SearchGene(gid="10", an.go=an.Hs.egGO, n=5)
}
