\name{LC}
\alias{LC}
\docType{data}
\title{LC}
\description{Level coefficent for all GO terms}
\usage{data(LC)}
\format{An environment}
\source{From function Onto2LC, for example,
#BPLC <- Onto2LC("BP")
#MFLC <- Onto2LC("MF")
#CCLC <- Onto2LC("CC")
#lc <- c(BPLC, MFLC, CCLC)
#LC <- new.env()
#for(i in 1:length(lc)){
#  assign(names(lc)[i], lc[i], envir=LC)
#}

}
\keyword{datasets}

