% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim.R
\name{simPricePath}
\alias{simPricePath}
\title{Simulate random walk path of stock price or else for given period}
\usage{
simPricePath(startPrice, volatility, days, expectedRR, daysInYear = 252)
}
\arguments{
\item{startPrice}{start price}

\item{volatility}{volatility}

\item{days}{number of days in path}

\item{expectedRR}{expected rate of return of stock or else}

\item{daysInYear}{number of business days in year}
}
\value{
vector of stock price path
}
\description{
Simulate random walk path of stock price or else for given period
}
